refactor: PassThroughPointsService out of api package
